Move RepresentativeTraversalFinder to genotypekit and pull in the functions it uses
