+++ src/call2vcf.cpp
@@ -159,14 +159,7 @@ double total(const FractionalSupport& support) {
 }
 
 /**
- * Get the total read support in a Support.
- */
-int total(const Support& support) {
-    return support.forward() + support.reverse();
-}
-
-/**
- * Get the strand bias of a support.
+ * Get the strand bias of a FractionalSupport.
  */
 double strand_bias(const FractionalSupport& support) {
     return std::max(support.first, support.second) / (support.first + support.second);
@@ -179,51 +172,6 @@ double strand_bias(const FractionalSupport& support) {
 FractionalSupport support_min(const FractionalSupport& a, const FractionalSupport& b) {
     return std::make_pair(std::min(a.first, b.first), std::min(a.second, b.second));
 }
-
-/**
- * Get the minimum support of a pair of Supports, by taking the min in each
- * orientation.
- */
-Support support_min(const Support& a, const Support& b) {
-    Support to_return;
-    to_return.set_forward(min(a.forward(), b.forward()));
-    to_return.set_reverse(min(a.reverse(), b.reverse()));
-    return to_return;
-}
-
-
-/**
- * This TraversalFinder is derived from the old vg call code, and emits at least
- * one traversal representing every node, and one traversal representing every
- * edge.
- */
-class RepresentativeTraversalFinder : public TraversalFinder {
-    
-    AugmentedGraph& augmented;
-    SnarlManager& snarl_manager;
-    PathIndex& index;
-    
-    // What DFS depth should we search to?
-    size_t max_depth;
-    // How many search intermediates can we allow?
-    size_t max_bubble_paths;
-    
-public:
-
-    RepresentativeTraversalFinder(AugmentedGraph& augmented, SnarlManager& snarl_manager,
-        PathIndex& index, size_t max_depth, size_t max_bubble_paths);
-    
-    bool verbose = false;
-    
-    virtual ~RepresentativeTraversalFinder() = default;
-    
-    /**
-     * Find traversals to cover the nodes and edges of the snarl. Always emits
-     * the primary path traversal first, if applicable.
-     */
-    virtual vector<SnarlTraversal> find_traversals(const Snarl& site);
-    
-};
 
 /**
  * Make a letter into a full string because apparently that's too fancy for the
@@ -293,704 +241,6 @@ bool mapping_is_perfect_match(const vg::Mapping& mapping) {
     // Note that Mappings with no edits at all are full-length perfect matches.
     return true;
 }
-
-/**
- * Get the length of a path through nodes, in base pairs.
- */
-size_t bp_length(const std::list<vg::NodeTraversal>& path) {
-    size_t length = 0;
-    for(auto& traversal : path) {
-        // Sum up length of each node's sequence
-        length += traversal.node->sequence().size();
-    }
-    return length;
-}
-
-/**
- * Get the minimum support of all nodes and edges in path
- */
-Support min_support_in_path(vg::VG& graph,
-                            const std::map<vg::Node*, Support>& node_supports,
-                            const std::map<vg::Edge*, Support>& edge_supports,
-                            const std::list<vg::NodeTraversal>& path) {
-    if (path.empty()) {
-        return Support();
-    }
-    auto cur = path.begin();
-    auto next = path.begin();
-    ++next;
-    Support minSupport = node_supports.count(cur->node) ? node_supports.at(cur->node) : Support();
-    for (; next != path.end(); ++cur, ++next) {
-        // check the node support
-        Support support = node_supports.count(next->node) ? node_supports.at(next->node) : Support();
-        minSupport = support_min(minSupport, support);
-        
-        // check the edge support
-        Edge* edge = graph.get_edge(*cur, *next);
-        assert(edge != NULL);
-        Support edgeSupport = edge_supports.count(edge) ? edge_supports.at(edge) : Support();
-        minSupport = support_min(minSupport, edgeSupport);
-    }
-
-    return minSupport;
-}
-
-/**
- * Do a breadth-first search left from the given node traversal, and return
- * lengths and paths starting at the given node and ending on the indexed
- * reference path. Refuses to visit nodes with no support.
- */
-std::set<std::pair<size_t, std::list<vg::NodeTraversal>>> bfs_left(vg::VG& graph,
-    vg::NodeTraversal node, const PathIndex& index,
-    const std::map<vg::Node*, Support>& node_supports,
-    const std::map<vg::Edge*, Support>& edge_supports,
-    int64_t maxDepth = 10, bool stopIfVisited = false) {
-
-    // Holds partial paths we want to return, with their lengths in bp.
-    std::set<std::pair<size_t, std::list<vg::NodeTraversal>>> toReturn;
-    
-    // Do a BFS
-    
-    // This holds the paths to get to NodeTraversals to visit (all of which will
-    // end with the node we're starting with).
-    std::list<std::list<vg::NodeTraversal>> toExtend;
-    
-    // This keeps a set of all the oriented nodes we already got to and don't
-    // need to queue again.
-    std::set<vg::NodeTraversal> alreadyQueued;
-    
-    // Start at this node at depth 0
-    toExtend.emplace_back(std::list<vg::NodeTraversal> {node});
-    // Mark this traversal as already queued
-    alreadyQueued.insert(node);
-    
-    // How many ticks have we spent searching?
-    size_t searchTicks = 0;
-    // Track how many options we have because size may be O(n).
-    size_t stillToExtend = toExtend.size();
-    
-    while(!toExtend.empty()) {
-        // Keep going until we've visited every node up to our max search depth.
-        
-#ifdef debug
-        searchTicks++;
-        if(searchTicks % 100 == 0) {
-            // Report on how much searching we are doing.
-            std::cerr << "Search tick " << searchTicks << ", " << stillToExtend << " options." << std::endl;
-        }
-#endif
-        
-        // Dequeue a path to extend.
-        // Make sure to move out of the list to avoid a useless copy.
-        std::list<vg::NodeTraversal> path(std::move(toExtend.front()));
-        toExtend.pop_front();
-        stillToExtend--;
-        
-        // We can't just throw out longer paths, because shorter paths may need
-        // to visit a node twice (in opposite orientations) and thus might get
-        // rejected later. Or they might overlap with paths on the other side.
-        
-        // Look up and see if the front node on the path is on our reference
-        // path
-        if(index.by_id.count(path.front().node->id())) {
-            // This node is on the reference path. TODO: we don't care if it
-            // lands in a place that is itself deleted.
-            
-            // Say we got to the right place
-            toReturn.emplace(bp_length(path), std::move(path));
-            
-            // Don't bother looking for extensions, we already got there.
-        } else if(path.size() <= maxDepth) {
-            // We haven't hit the reference path yet, but we also haven't hit
-            // the max depth. Extend with all the possible extensions.
-            
-            // Look left
-            vector<vg::NodeTraversal> prevNodes;
-            graph.nodes_prev(path.front(), prevNodes);
-            
-            for(auto prevNode : prevNodes) {
-                // For each node we can get to
-                vg::Edge* edge = graph.get_edge(prevNode, path.front());
-                assert(edge != NULL);
-                
-                if((!node_supports.empty() && (!node_supports.count(prevNode.node) ||
-                    total(node_supports.at(prevNode.node)) == 0)) ||
-                   (!edge_supports.empty() && (!edge_supports.count(edge) ||
-                    total(edge_supports.at(edge)) == 0))) {
-                    
-                    // We have no support at all for visiting this node (but we
-                    // do have some node read support data)
-                    continue;
-                }
-                
-                if(stopIfVisited && alreadyQueued.count(prevNode)) {
-                    // We already have a way to get here.
-                    continue;
-                }
-            
-                // Make a new path extended left with the node
-                std::list<vg::NodeTraversal> extended(path);
-                extended.push_front(prevNode);
-                toExtend.emplace_back(std::move(extended));
-                stillToExtend++;
-                
-                // Remember we found a way to this node, so we don't try and
-                // visit it other ways.
-                alreadyQueued.insert(prevNode);
-            }
-        }
-        
-    }
-    
-    return toReturn;
-}
-
-/**
- * Flip a NodeTraversal around and return the flipped copy.
- */
-vg::NodeTraversal flip(vg::NodeTraversal toFlip) {
-    return vg::NodeTraversal(toFlip.node, !toFlip.backward);
-}
-
-/**
- * Do a breadth-first search right from the given node traversal, and return
- * lengths and paths starting at the given node and ending on the indexed
- * reference path.
- */
-std::set<std::pair<size_t, std::list<vg::NodeTraversal>>> bfs_right(vg::VG& graph,
-    vg::NodeTraversal node, const PathIndex& index,
-    const std::map<vg::Node*, Support>& node_supports,
-    const std::map<vg::Edge*, Support>& edge_supports,
-    int64_t maxDepth = 10, bool stopIfVisited = false) {
-
-    // Look left from the backward version of the node.
-    auto toConvert = bfs_left(graph, flip(node), index, node_supports,
-                              edge_supports, maxDepth, stopIfVisited);
-    
-    // Since we can't modify set records in place, we need to do a copy
-    std::set<std::pair<size_t, std::list<vg::NodeTraversal>>> toReturn;
-    
-    for(auto lengthAndPath : toConvert) {
-        // Flip every path to run the other way
-        lengthAndPath.second.reverse();
-        for(auto& traversal : lengthAndPath.second) {
-            // And invert the orientation of every node in the path in place.
-            traversal = flip(traversal);
-        }
-        // Stick it in the new set
-        toReturn.emplace(std::move(lengthAndPath));
-    }
-    
-    return toReturn;
-}
-
-
-/**
- * Given a vg graph, an edge or node in the graph, and an index for the
- * reference path, look out from the edge or node in both directions to find a
- * shortest bubble relative to the path, with a consistent orientation. The
- * bubble may not visit the same node twice.
- *
- * Exactly one of edge and node must be null, and one not null.
- *
- * Takes a max depth for the searches producing the paths on each side.
- * 
- * Return the ordered and oriented nodes in the bubble, with the outer nodes
- * being oriented forward along the named path, and with the first node coming
- * before the last node in the reference.  Also return the minimum support
- * found on any edge or node in the bubble (including the reference node endpoints
- * and their edges which aren't stored in the path)
- */
-std::pair<Support, std::vector<vg::NodeTraversal> >
-find_bubble(vg::VG& graph, vg::Node* node, vg::Edge* edge, const PathIndex& index,
-            const std::map<vg::Node*, Support>& node_supports,
-            const std::map<vg::Edge*, Support>& edge_supports, int64_t maxDepth,
-            size_t max_bubble_paths) {
-
-    // What are we going to find our left and right path halves based on?
-    NodeTraversal left_traversal;
-    NodeTraversal right_traversal;
-
-    if(edge != nullptr) {
-        // Be edge-based
-        
-        // Find the nodes at the ends of the edges. Look at them traversed in the
-        // edge's local orientation.
-        left_traversal = NodeTraversal(graph.get_node(edge->from()), edge->from_start());
-        right_traversal = NodeTraversal(graph.get_node(edge->to()), edge->to_end());
-        
-    } else {
-        // Be node-based
-        assert(node != nullptr);
-        left_traversal = right_traversal = NodeTraversal(node);
-    }
-    
-#ifdef debug
-    std::cerr << "Starting from: " << left_traversal << ", " << right_traversal << std::endl;
-#endif
-
-    // Find paths on both sides, with nodes on the primary path at the outsides
-    // and this edge in the middle. Returns path lengths and paths in pairs in a
-    // set.
-    auto leftPaths = bfs_left(graph, left_traversal, index, node_supports,
-                              edge_supports, maxDepth);
-    auto rightPaths = bfs_right(graph, right_traversal, index, node_supports,
-                                edge_supports, maxDepth);
-    
-    // Find a combination of two paths which gets us to the reference in a
-    // consistent orientation (meaning that when you look at the ending nodes'
-    // Mappings in the reference path, the ones with minimal ranks have the same
-    // orientations) and which doesn't use the same nodes on both sides.
-    // Track support of up to max_bubble_paths combinations, and return the
-    // highest
-    std::pair<Support, std::vector<NodeTraversal> > bestBubblePath;
-    int bubbleCount = 0;
-    
-    // We need to look in different combinations of lists.
-    auto testCombinations = [&](const std::list<std::list<vg::NodeTraversal>>& leftList,
-        const std::list<std::list<vg::NodeTraversal>>& rightList) {
-
-        for(auto leftPath : leftList) {
-            // Figure out the relative orientation for the leftmost node.
-#ifdef debug        
-            std::cerr << "Left path: " << std::endl;
-            for(auto traversal : leftPath ) {
-                std::cerr << "\t" << traversal << std::endl;
-            }
-#endif    
-            // Split out its node pointer and orientation
-            auto leftNode = leftPath.front().node;
-            auto leftOrientation = leftPath.front().backward;
-            
-            // Get where it falls in the reference as a position, orientation pair.
-            auto leftRefPos = index.by_id.at(leftNode->id());
-            
-            // We have a backward orientation relative to the reference path if we
-            // were traversing the anchoring node backwards, xor if it is backwards
-            // in the reference path.
-            bool leftRelativeOrientation = leftOrientation != leftRefPos.second;
-            
-            // Make a set of all the nodes in the left path
-            std::set<int64_t> leftPathNodes;
-            for(auto visit : leftPath) {
-                leftPathNodes.insert(visit.node->id());
-            }
-
-            // Get the minimum support in the left path
-            Support minLeftSupport = min_support_in_path(
-                graph, node_supports, edge_supports, leftPath);
-            
-            for(auto rightPath : rightList) {
-                // Figure out the relative orientation for the rightmost node.
-#ifdef debug            
-                std::cerr << "Right path: " << std::endl;
-                for(auto traversal : rightPath ) {
-                    std::cerr << "\t" << traversal << std::endl;
-                }
-#endif            
-                // Split out its node pointer and orientation
-                // Remember it's at the end of this path.
-                auto rightNode = rightPath.back().node;
-                auto rightOrientation = rightPath.back().backward;
-                
-                // Get where it falls in the reference as a position, orientation pair.
-                auto rightRefPos = index.by_id.at(rightNode->id());
-                
-                // We have a backward orientation relative to the reference path if we
-                // were traversing the anchoring node backwards, xor if it is backwards
-                // in the reference path.
-                bool rightRelativeOrientation = rightOrientation != rightRefPos.second;
-
-                // Get the minimum support in the right path
-                Support minRightSupport = min_support_in_path(
-                    graph, node_supports, edge_supports, rightPath);
-                
-                if(leftRelativeOrientation == rightRelativeOrientation &&
-                    ((!leftRelativeOrientation && leftRefPos.first < rightRefPos.first) ||
-                    (leftRelativeOrientation && leftRefPos.first > rightRefPos.first))) {
-                    // We found a pair of paths that get us to and from the
-                    // reference without turning around, and that don't go back to
-                    // the reference before they leave.
-
-                    // Get the minimum support of combined left and right paths
-                    Support minFullSupport = support_min(minLeftSupport, minRightSupport);
-                    
-                    // Start with the left path
-                    std::vector<vg::NodeTraversal> fullPath{leftPath.begin(), leftPath.end()};
-                    
-                    // We need to detect overlap with the left path
-                    bool overlap = false;
-                    
-                    // If we're starting from an edge, we keep the first node on
-                    // the right path. If we're starting from a node, we need to
-                    // discard it because it's just another copy of our node
-                    // we're starting with.
-                    for(auto it = (edge != nullptr ? rightPath.begin() : ++(rightPath.begin())); it != rightPath.end(); ++it) {
-                        // For all but the first node on the right path, add that in
-                        fullPath.push_back(*it);
-                        
-                        if(leftPathNodes.count((*it).node->id())) {
-                            // We already visited this node on the left side. Try
-                            // the next right path instead.
-                            overlap = true;
-                        }
-                    }
-                    
-                    if(overlap) {
-                        // Can't combine this right with this left, as they share
-                        // nodes and we can't handle the copy number implications.
-                        // Try the next right.
-                        // TODO: handle the copy number implications.
-                        continue;
-                    }
-                    
-                    if(leftRelativeOrientation) {
-                        // Turns out our anchored path is backwards.
-                        
-                        // Reorder everything the other way
-                        std::reverse(fullPath.begin(), fullPath.end());
-                        
-                        for(auto& traversal : fullPath) {
-                            // Flip each traversal
-                            traversal = flip(traversal);
-                        }
-                    }
-
-                    
-#ifdef debug        
-                    std::cerr << "Merged path:" << std::endl;
-                    for(auto traversal : fullPath) {
-                        std::cerr << "\t" << traversal << std::endl;
-                    }                    
-#endif
-                    // update our best path by seeing if we've found one with higher min support
-                    if (total(minFullSupport) > total(bestBubblePath.first)) {
-                        bestBubblePath.first = minFullSupport;
-                        bestBubblePath.second = fullPath;
-                    }
-
-                    // keep things from getting out of hand
-                    if (++bubbleCount >= max_bubble_paths) {
-                        return bestBubblePath;
-                    }
-                }
-            }
-        }
-        
-        // Return the best path along with its min support
-        // (could be empty)
-        return bestBubblePath;
-        
-    };
-    
-    // Convert sets to lists, which requires a copy again...
-    std::list<std::list<vg::NodeTraversal>> leftConverted;
-    for(auto lengthAndPath : leftPaths) {
-        leftConverted.emplace_back(std::move(lengthAndPath.second));
-    }
-    std::list<std::list<vg::NodeTraversal>> rightConverted;
-    for(auto lengthAndPath : rightPaths) {
-        rightConverted.emplace_back(std::move(lengthAndPath.second));
-    }
-    
-    // Look for a valid combination, or return an empty path if one iesn't
-    // found.
-    return testCombinations(leftConverted, rightConverted);
-    
-}
-
-RepresentativeTraversalFinder::RepresentativeTraversalFinder(AugmentedGraph& augmented,
-    SnarlManager& snarl_manager, PathIndex& index, size_t max_depth,
-    size_t max_bubble_paths) : augmented(augmented), snarl_manager(snarl_manager),
-    index(index), max_depth(max_depth), max_bubble_paths(max_bubble_paths) {
-    
-    // Nothing to do!
-
-}
-
-vector<SnarlTraversal> RepresentativeTraversalFinder::find_traversals(const Snarl& site) {
-    
-    // TODO: we can only do ultrabubbles right now. Other snarls may not have
-    // traversals through from end to end.
-    assert(site.type() == ULTRABUBBLE);
-    
-    // Get its nodes and edges (including all child sites)
-    pair<unordered_set<Node*>, unordered_set<Edge*>> contents = snarl_manager.deep_contents(&site, augmented.graph, true);
-    
-    // Copy its node set
-    std::unordered_set<Node*> nodes_left(contents.first);
-
-    // Trace the ref path through the site
-    std::vector<NodeTraversal> ref_path_for_site;
-    
-    // First figure where the site starts and ends in the primary path
-    // TODO: support off-path sites
-    size_t site_start = index.by_id.at(site.start().node_id()).first;
-    size_t site_end = index.by_id.at(site.end().node_id()).first;
-    
-    
-#ifdef debug
-    std::cerr << "Site starts with " << to_node_traversal(site.start(), augmented.graph)
-        << " at " << site_start
-        << " and ends with " << to_node_traversal(site.end(), augmented.graph)
-        << " at " << site_end << std::endl;
-#endif
-
-    // The primary path may go through the site backward. So get the primary min and max coords
-    size_t primary_min = min(site_start, site_end);
-    size_t primary_max = max(site_start, site_end);
-
-    // Then walk nodes from min coordinate to max coordinate. This holds the
-    // start coordinate of the current node.
-    int64_t ref_node_start = primary_min;
-    while(ref_node_start <= primary_max) {
-    
-        // Find the reference node starting here or later.
-        auto found = index.by_start.lower_bound(ref_node_start);
-        if(found == index.by_start.end()) {
-            throw runtime_error("No ref node found when tracing through site!");
-        }
-        if((*found).first > index.by_id.at(site.end().node_id()).first) {
-            // The next reference node we can find is out of the space
-            // being replaced. We're done.
-            if (verbose) {
-                cerr << "Stopping for out-of-bounds node" << endl;
-            }
-            break;
-        }
-        
-        // Get the corresponding NodeTraversal (by converting through Visit)
-        NodeTraversal found_traversal = to_node_traversal(found->second.to_visit(), augmented.graph);
-        
-        // Add the traversal to the ref path through the site
-        ref_path_for_site.push_back(found_traversal);
-        
-        // Make sure this node is actually in the site
-        assert(contents.first.count(found_traversal.node));
-        
-        // Drop it from the set of nodes in the site we haven't visited.
-        nodes_left.erase(found_traversal.node);
-        
-        // Next iteration look where this node ends.
-        ref_node_start = found->first + found_traversal.node->sequence().size();
-    }
-    
-    for(auto node : nodes_left) {
-        // Make sure none of the nodes in the site that we didn't visit
-        // while tracing along the ref path are on the ref path.
-        if(index.by_id.count(node->id())) {
-            std::cerr << "Node " << node->id() << " is on ref path at "
-                << index.by_id.at(node->id()).first << " but not traced in site "
-                << to_node_traversal(site.start(), augmented.graph) << " to " 
-                << to_node_traversal(site.end(), augmented.graph) << std::endl;
-            throw runtime_error("Extra ref node found");
-        }
-    }
-    
-    // We need to know all the full-length traversals we're going to consider.
-    // XREF states will have to be calculated later, over the whole traversal.
-    std::set<std::vector<NodeTraversal>> site_traversal_set;
-    
-    // We have this function to extend a partial traversal into a full
-    // traversal and add it as a path. The path must already be rooted on
-    // the reference in the correct order and orientation.
-    auto extend_into_allele = [&](std::vector<NodeTraversal> path) {
-        // Splice the ref path through the site and the bubble's path
-        // through the site together.
-        vector<NodeTraversal> extended_path;
-
-        for(auto& traversal : path) {
-            // Make sure the site actually has the nodes we're visiting.
-            assert(contents.first.count(traversal.node));
-#ifdef debug
-            if(index.by_id.count(traversal.node->id())) {
-                std::cerr << "Path member " << traversal << " lives on ref at "
-                << index.by_id.at(traversal.node->id()).first << std::endl;
-            } else {
-                std::cerr << "Path member " << traversal << " does not live on ref" << std::endl;
-            }
-#endif
-        }
-        
-        size_t ref_path_index = 0;
-        size_t bubble_path_index = 0;
-        
-        while(ref_path_for_site.at(ref_path_index) != path.at(bubble_path_index)) {
-            // Collect NodeTraversals from the ref path until we hit the one
-            // at which the bubble path starts.
-            extended_path.push_back(ref_path_for_site[ref_path_index++]);
-        }
-        while(bubble_path_index < path.size()) {
-            // Then take the whole bubble path
-            extended_path.push_back(path[bubble_path_index++]);
-        }
-        while(ref_path_index < ref_path_for_site.size() && ref_path_for_site.at(ref_path_index) != path.back()) {
-            // Then skip ahead to the matching point in the ref path
-            ref_path_index++;
-        }
-        if(ref_path_index == ref_path_for_site.size()) {
-            // We ran out of ref path before finding the place to leave the alt.
-            // This must be a backtracking loop or something; start over from the beginning.
-            ref_path_index = 0;
-            
-            while(ref_path_index < ref_path_for_site.size() && ref_path_for_site.at(ref_path_index) != path.back()) {
-                // Then skip ahead to the matching point in the ref path
-                ref_path_index++;
-            }
-            
-            if(ref_path_index == ref_path_for_site.size()) {
-                // Still couldn't find it!
-                std::stringstream err;
-                err << "Couldn't find " << path.back() << " in ref path of site "
-                    << to_node_traversal(site.start(), augmented.graph)
-                    << " to " << to_node_traversal(site.end(), augmented.graph) << endl;
-                throw std::runtime_error(err.str());
-            }
-        }
-        // Skip the matching NodeTraversal
-        ref_path_index++;
-        while(ref_path_index < ref_path_for_site.size()) {
-            // Then take the entier rest of the ref path
-            extended_path.push_back(ref_path_for_site[ref_path_index++]);
-        }
-        
-        // Now add it to the set
-        site_traversal_set.insert(extended_path);
-    
-    };
-
-    for(Node* node : contents.first) {
-        // Find the bubble for each node
-        
-        if(total(augmented.node_supports.at(node)) == 0) {
-            // Don't bother with unsupported nodes
-            continue;
-        }
-        
-        if(index.by_id.count(node->id())) {
-            // Don't try to pathfind to the reference for reference nodes.
-            continue;
-        }
-        
-        // TODO: allow bubbles that don't backend into the primary path
-        std::pair<Support, std::vector<NodeTraversal>> sup_path = find_bubble(
-            augmented.graph, node, nullptr, index, augmented.node_supports,
-            augmented.edge_supports, max_depth, max_bubble_paths);
-
-        std::vector<NodeTraversal>& path = sup_path.second;
-        
-        if(path.empty()) {
-            // We couldn't find a path back to the primary path. Discard
-            // this material.
-            if (verbose) {
-                cerr << "Warning: No path found for node " << node->id() << endl;
-            }
-            // TODO: record the node's bases as lost.
-            
-            // TODO: what if it's already in another bubble/the node is deleted?
-            continue;
-        }
-        
-        // Extend it out into an allele
-        extend_into_allele(path);
-        
-    }
-    
-    for(Edge* edge : contents.second) {
-        // Go through all the edges
-        
-        if(!index.by_id.count(edge->from()) || !index.by_id.count(edge->to())) {
-            // Edge doesn't touch reference at both ends. Don't use it
-            // because for some reason it makes performance worse
-            // overall.
-            continue;
-        }
-        
-        // Find a path based around this edge
-        std::pair<Support, std::vector<NodeTraversal>> sup_path = find_bubble(
-            augmented.graph, nullptr, edge, index, augmented.node_supports,
-            augmented.edge_supports, max_depth, max_bubble_paths);
-        std::vector<NodeTraversal>& path = sup_path.second;
-        
-#ifdef debug
-        std::cerr << "Edge " << edge->from() << " to " << edge->to() << " yields:" << std::endl;
-        for(auto& traversal : path) {
-            std::cerr << "\t" << traversal << std::endl;
-        }
-#endif
-        
-        if(path.empty()) {
-            // We couldn't find a path back to the primary path. Discard
-            // this material.
-            if (verbose) {
-                cerr << "Warning: No path found for edge " << edge->from() << "," << edge->to() << endl;
-            }
-            // TODO: bases lost
-            // TODO: what if it's already in another bubble/the node is deleted?
-            continue;
-        }
-        
-        // Extend it out into an allele
-        extend_into_allele(path);
-    }
-    
-    
-    // Now convert to SnarlTraversals
-    vector<SnarlTraversal> unique_traversals;
-    
-    // Have a function to convert a vector of NodeTraversals including the snarl
-    // ends into a SnarlTraversal
-    auto emit_traversal = [&](vector<NodeTraversal> node_traversals) {
-        // Make it into this SnarlTraversal
-        SnarlTraversal trav;
-        
-        // Label traversal with the snarl
-        *trav.mutable_snarl()->mutable_start() = site.start();
-        *trav.mutable_snarl()->mutable_end() = site.end();
-        
-        // Add everything but the first and last nodes as Visits.
-        // TODO: think about nested sites?
-        
-        if (site_start > site_end) {
-            // The primary path runs backward, so we need to emit all our lists
-            // of NodeTraversals backward so they come out in the snarl's
-            // orientation and not the primary path's.
-            
-            for(size_t i = 1; i + 1 < node_traversals.size(); i++) {
-                // Make a Visit for each NodeTraversal but the first and last,
-                // but backward and in reverse order.
-                *trav.add_visits() = to_visit(node_traversals[node_traversals.size() - 1 - i].reverse());
-            }
-            
-        } else {
-            // The primary path and the snarl use the same orientation
-            
-            for(size_t i = 1; i + 1 < node_traversals.size(); i++) {
-                // Make a Visit for each NodeTraversal but the first and last
-                *trav.add_visits() = to_visit(node_traversals[i]);
-            }
-            
-        }
-        
-        // Save the SnarlTraversal
-        unique_traversals.push_back(trav);
-    };
-    
-    
-    // Do the ref path first
-    emit_traversal(ref_path_for_site);
-    for(auto& node_traversals : site_traversal_set) {
-        // Look at each vector of NodeTraversals
-        if (node_traversals != ref_path_for_site) {
-            // And do everything other than the ref path
-            emit_traversal(node_traversals);            
-        }
-        
-    }
-    
-    return unique_traversals;
-}
-
 
 /**
  * Given a collection of pileups by original node ID, and a set of original node
